+++ jdaviz/configs/cubeviz/plugins/tests/test_data_retrieval.py
@@ -29,9 +29,13 @@ def test_data_retrieval(cubeviz_helper):
     # two ways of retrieving data from the viewer.
     # They should return the same spectral values
     a1 = cubeviz_helper.app.get_viewer(spectrum_viewer_reference_name).data()
-    a2 = cubeviz_helper.get_data("Spectrum (sum)")
+    a2 = cubeviz_helper.datasets["Spectrum (sum)"].get_data()
+    # Test the old API as well
+    a3 = cubeviz_helper.get_data("Spectrum (sum)")
 
     test_value_1 = a1[0].data
     test_value_2 = a2.flux.value
+    test_value_3 = a3.flux.value
 
     assert np.allclose(test_value_1, test_value_2, atol=1e-5)
+    assert np.allclose(test_value_2, test_value_3, atol=1e-5)