+++ jdaviz/configs/specviz2d/tests/test_parsers.py
@@ -82,7 +82,7 @@ def test_hlsp_goods_s2d_deconfigged(deconfigged_helper):
     dc_0 = deconfigged_helper.app.data_collection[0]
     assert dc_0.get_component('flux').shape == (27, 674)
     assert isinstance(deconfigged_helper.plugins['2D Spectral Extraction'].trace_dataset.selected_obj, Spectrum)  # noqa
-    assert isinstance(deconfigged_helper.get_data('2D Spectrum'), Spectrum)
+    assert isinstance(deconfigged_helper.datasets['2D Spectrum'].get_data(), Spectrum)
 
 
 def test_load_two_2d_spectra_deconfigged(deconfigged_helper):