+++ CHANGES.rst
@@ -25,6 +25,8 @@ New Features
 
 - Indicate in loaders whether the loaded entry/entries will overwrite existing data in the app. [#3997]
 
+- Top level access to datasets API dictionary, deprecating top-level access to data_labels list. [#4024]
+
 Cubeviz
 ^^^^^^^